proxy app install from conda working extracted mamba commands into own file resolved circulating references
