+++ tethysapp/app_store/application_files/proxyapp_setup_template.py
@@ -1,6 +1,10 @@
-from setuptools import setup
+from setuptools import setup, find_packages
 
 setup(
     name="$app_name",
-    version="$app_version"
+    version="$app_version",
+    packages=find_packages(),
+    package_data={"": ["*.yml", "*.yaml"]},
+    include_package_data=True,
+    zip_safe=False
 )
