+++ tethysapp/app_store/helpers.py
@@ -299,3 +299,81 @@ def get_color_label_dict(stores):
             index_style += 1
 
     return color_store_dict, stores
+
+
+def clear_github_cache_list():
+    """Clears out the stored cache of GitHub installed apps.
+    """
+    cache.delete(CACHE_KEY)
+
+
+def restart_server(data, channel_layer, app_workspace, run_collect_all=True):
+    """Runs some tethys commands after an application is installed. Once finished, try to restart the server to get
+    the changes made
+
+    Args:
+        data (dict): Dictionary of data with app information and restart type
+        channel_layer (Django Channels Layer): Asynchronous Django channel layer from the websocket consumer
+        app_workspace (TethysWorkspace): workspace object bound to the app workspace.
+        run_collect_all (bool, optional): Detemines if collect all needs to be ran. Defaults to True.
+    """
+    # Check if Install Running file is present and delete it
+    workspace_directory = app_workspace.path
+    install_running_path = os.path.join(workspace_directory, 'install_status', 'installRunning')
+    if os.path.exists(install_running_path):
+        os.remove(install_running_path)
+
+    # Check if Scaffold Running file is present and delete it
+    scaffold_running_path = os.path.join(workspace_directory, 'install_status', 'scaffoldRunning')
+    if os.path.exists(scaffold_running_path):
+        os.remove(scaffold_running_path)
+
+    manage_path = get_manage_path({})
+    if data["restart_type"] in ["install", "update", "github_install", "scaffold_install"]:
+        # Run SyncStores
+        logger.info("Running Syncstores for app: " + data["name"])
+        send_notification("Running Syncstores for app: " + data["name"], channel_layer)
+        intermediate_process = ['python', manage_path, 'syncstores', data["name"], '-f']
+        run_process(intermediate_process)
+
+    if 'runserver' in sys.argv:
+
+        logger.info("Dev Mode. Attempting to restart by changing file")
+        dir_path = os.path.dirname(os.path.realpath(__file__))
+        file_path = os.path.join(dir_path, 'model.py')
+        with open(file_path, "w") as f:
+            f.write(f'print("{data["name"]} installed in dev mode")')
+            f.write("\n")
+    else:
+        if run_collect_all and data["restart_type"] in ["install", "update", "github_install", "scaffold_install"]:
+
+            logger.info("Running Tethys Collectall")
+            send_notification("Running Tethys Collectall for app: " + data["name"], channel_layer)
+            intermediate_process = ['python', manage_path, 'pre_collectstatic']
+            run_process(intermediate_process)
+            # Setup for main collectstatic
+            intermediate_process = ['python', manage_path, 'collectstatic', '--noinput']
+            run_process(intermediate_process)
+            # Run collectworkspaces command
+            intermediate_process = ['python', manage_path, 'collectworkspaces', '--force']
+            run_process(intermediate_process)
+
+        try:
+            send_notification("Server Restarting . . .", channel_layer)
+            command = 'supervisorctl restart all'
+            subprocess.run(['sudo', '-h'], check=True)
+            sudoPassword = app.get_custom_setting('sudo_server_pass')
+
+            os.system('echo %s|sudo -S %s' % (sudoPassword, command))
+        except Exception as e:
+            logger.error(e)
+            logger.info("No SUDO. Docker container implied. Restarting without SUDO")
+            # Error encountered while running sudo. Let's try without sudo
+            # Check if the restart dir exists. If yes then use that dir instead of
+            RESTART_FILE_PATH = "/var/lib/tethys_persist/restart"
+
+            if os.path.isdir(RESTART_FILE_PATH):
+                logger.info("Restart Directory found. Creating restart file.")
+                Path(os.path.join(RESTART_FILE_PATH, 'restart')).touch()
+            else:
+                os.system(command)
