+++ tethysapp/app_store/templates/app_store/incompatibleAppsTable.html
@@ -8,8 +8,8 @@
                     <thead>
                         <tr>
                             <th data-field="name" data-formatter="mergedNameFormatter">App Name</th>
+                            <th data-field="app_type" data-formatter="mergedTypeFormatter">App Type</th>
                             <th data-field="latestVersion" data-formatter="mergedFieldsFormatter">App Latest Version</th>
-                            <!-- <th data-field="author" data-formatter="mergedFieldsFormatter" >Developer</th> -->
                             <th data-field="installed" data-formatter="mergedOperateFormatter" data-events="operateEvents">Actions</th>
                         </tr>
                     </thead>
