+++ tethysapp/app_store/templates/app_store/installedAppsTable.html
@@ -8,6 +8,7 @@
                     <thead>
                         <tr>
                             <th data-field="name" data-formatter="mergedNameFormatter">App Name</th>
+                            <th data-field="app_type" data-formatter="mergedTypeFormatter">App Type</th>
                             <th data-field="installedVersion" data-formatter="mergedFieldsFormatter">Installed Version</th>
                             <th data-field="latestVersion" data-formatter="mergedFieldsFormatter">Latest Version</th>
                             <th data-field="installed" data-formatter="mergedOperateFormatter" data-events="operateEvents">Actions</th>
